+++ zpa/resource_zpa_policy_capabilities_access_rule.go
@@ -47,7 +47,7 @@ func resourcePolicyCapabilitiesAccessRule() *schema.Resource {
 				Computed: true,
 			},
 			"conditions": {
-				Type:        schema.TypeList,
+				Type:        schema.TypeSet,
 				Optional:    true,
 				Computed:    true,
 				Description: "This is for proviidng the set of conditions for the policy.",
@@ -67,7 +67,7 @@ func resourcePolicyCapabilitiesAccessRule() *schema.Resource {
 							}, false),
 						},
 						"operands": {
-							Type:        schema.TypeList,
+							Type:        schema.TypeSet,
 							Optional:    true,
 							Computed:    true,
 							Description: "This signifies the various policy criteria.",
@@ -95,7 +95,7 @@ func resourcePolicyCapabilitiesAccessRule() *schema.Resource {
 										}, false),
 									},
 									"entry_values": {
-										Type:     schema.TypeList,
+										Type:     schema.TypeSet,
 										Optional: true,
 										Computed: true,
 										Elem: &schema.Resource{