+++ zpa/resource_zpa_policy_capabilities_access_rule_test.go
@@ -46,12 +46,12 @@ func TestAccResourcePolicyCapabilitiesAccessRule_Basic(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceTypeAndName, "conditions.#", "1"),
 				),
 			},
-			// Import test
-			// {
-			// 	ResourceName:      resourceTypeAndName,
-			// 	ImportState:       true,
-			// 	ImportStateVerify: true,
-			// },
+			//Import test
+			{
+				ResourceName:      resourceTypeAndName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
 		},
 	})
 }