+++ docs/guides/release-notes.md
@@ -12,10 +12,21 @@ Track all ZPA Terraform provider's releases. New resources, features, and bug fi
 
 ---
 
-``Last updated: v3.33.1``
+``Last updated: v3.33.2``
 
 ---
 
+## 3.33.2 (September, 10 2024)
+
+### Notes
+
+- Release date: **(September, 10 2024)**
+- Supported Terraform version: **v1.x**
+
+### Bug Fixes
+- [PR #486](https://github.com/zscaler/terraform-provider-zpa/pull/486) - Fixed drift related to common flattening and expand functions for all v2 Access policy resources.
+  ~> **NOTE** This fix does not affect existing configurations using the `v2` policy type.
+  
 ## 3.33.1 (September, 8 2024)
 
 ### Notes